requirements.txt
ops >= 1.2.0
pyyaml